+++ internal/model/like_types.go
@@ -10,7 +10,7 @@ type Like struct {
 	ID           primitive.ObjectID `bson:"_id,omitempty" json:"id,omitempty"`
 	UserId       string             `bson:"userId,omitempty" json:"userId,omitempty"`
 	TargetId     string             `bson:"targetId,omitempty" json:"targetId,omitempty"`
-	TargetType   int                `bson:"targetType,omitempty" json:"targetType,omitempty"`
+	TargetType   int64              `bson:"targetType,omitempty" json:"targetType,omitempty"`
 	AssociatedId string             `bson:"associatedId,omitempty" json:"associatedId,omitempty"`
 	UpdateAt     time.Time          `bson:"updateAt,omitempty" json:"updateAt,omitempty"`
 	CreateAt     time.Time          `bson:"createAt,omitempty" json:"createAt,omitempty"`
